requirements.txt
flask
flask-cors
gradio>=4.38.0
pillow>=10.4.0
pypdf>=4.2.0
pdf2image>=1.17.0
pytesseract>=0.3.10
qrcode>=7.4.2
pyzbar>=0.1.9
pydub>=0.25.1
python-dotenv>=1.0.1
werkzeug
psycopg2-binary
opencv-python
gunicorn
